upload files
Stockara is an application for tracking value of merchandise in stock.

It reads data from the database, or overwrites it with new data from uploaded csv file. It displays a custom svg chart with total values based on selected categories. Only library used is jQuery.

Date: 2017-08-29

Contact: Rokin - rokin3@gmail.com